# swaggerui/openapi.yaml
openapi: 3.0.0
info:
  description: Network Configurator service
  version: 0.3.0
  title: netConfD API
  contact:
    email: support@athonet.com
servers:
  - url: /api/1
    description: devmode
  - url: 'http://{server}/{prefix}/api/1'
    description: production
    variables:
      server:
        default: ngcore
        description: Server IP / Hostname
      prefix:
        default: ''
        description: Service platform prefix
  - url: 'https://{server}/{prefix}/api/1'
    description: production
    variables:
      server:
        default: ngcore
        description: Server IP / Hostname
      prefix:
        default: ''
        description: Service platform prefix
...
paths:
...
  /network/links:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigLinksGet
      summary: |
        Get All Links
      description: |
        Returns the array of the existing link layer interfaces by their ifname id
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/link'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigLinkCreate
      summary: |
        Create New Link
      description: |
        Configures and brings up a link layer interface.
        If the link exists return an error
      requestBody:
        description: Link Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/link'
      responses:
        '201':
          description: Successful creation
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
  /network/links/{ifname}:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigLinkGet
      summary: |
        Get Link
      description: |
        Retrieve link layer interface information
      parameters:
      - in: path
        name: ifname
        required: true
        schema:
          $ref: '#/components/schemas/ifname'
      responses:
        '200':
          description: Successful get
        "404":
          description: Not found
    delete:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigLinkDel
      summary: |
        Delete Link
      description: |
        Brings down and delete a link layer interface

        Only virtual interfaces can be deleted; physical (`eth*`) interfaces
        exist throughout the whole runtime of the system and cannot be deleted.
      parameters:
      - in: path
        name: ifname
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful delete
        "404":
          description: Not found
  /network/routes:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigRoutesGet
      summary: |
        Get All Routes
      description: |
        Returns the array of the existing routing tables routes ids
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/route'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigRouteCreate
      summary: |
        Create New Route
      description: |
        Configures a L3 route on the main routing table
      requestBody:
        description: Route Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/route'
      responses:
        '201':
          description: Successful creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/routeid'
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
  /network/routes/{routeid}:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigRouteGet
      summary: |
        Get Route
      description: |
        Get a L3 route details
      parameters:
      - in: path
        name: routeid
        required: true
        schema:
          $ref: '#/components/schemas/routeid'
      responses:
        '200':
          description: Successful get
        "404":
          description: Not found
    delete:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigRouteDel
      summary: |
        Delete Route
      description: |
        Brings down and delete an L3 IP route
      parameters:
      - in: path
        name: routeid
        required: true
        schema:
          $ref: '#/components/schemas/routeid'
      responses:
        '200':
          description: Successful delete
        "404":
          description: Not found
  /network/dhcp:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigDHCPsGet
      summary: |
        Get All DHCP
      description: |
        Returns the array of the existing link layer interfaces 
        controlled via DHCP and related config/state
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/dhcp'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigDHCPCreate
      summary: |
        Create DHCP
      description: |
        Adds a link interface under DHCP management
      requestBody:
        description: DHCP Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dhcp'
      responses:
        '201':
          description: Successful creation
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
  /network/dhcp/{ifname}:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigDHCPGet
      summary: |
        Get DHCP
      description: |
        Retrieve DHCP info for link layer interface
      parameters:
      - in: path
        name: ifname
        required: true
        schema:
          $ref: '#/components/schemas/ifname'
      responses:
        '200':
          description: Successful get
        "404":
          description: Not found
    delete:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigDHCPDel
      summary: |
        Delete DHCP
      description: |
        Delete DHCP daemon management control for the 
        specified link layer interface
      parameters:
      - in: path
        name: ifname
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful delete
        "404":
          description: Not found
  /network/dns:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigDNSsGet
      summary: |
        Get All DNS config
      description: |
        Returns the array of the existing nameservers for DNS queries.
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/dns'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigDNSCreate
      summary: |
        Create DNS
      description: |
        Adds a nameserver entry
      requestBody:
        description: DNS Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dns'
      responses:
        '201':
          description: Successful creation
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
  /network/dns/{dnsid}:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigDNSGet
      summary: |
        Get DNS
      description: |
        Retrieve DNS info for link layer interface
      parameters:
      - in: path
        name: dnsid
        required: true
        schema:
          $ref: '#/components/schemas/dnsid'
      responses:
        '200':
          description: Successful get
        "404":
          description: Not found
    delete:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigDNSDel
      summary: |
        Delete DNS
      description: |
        Delete DNS daemon management control for the 
        specified link layer interface
      parameters:
      - in: path
        name: dnsid
        required: true
        schema:
          $ref: '#/components/schemas/dnsid'
      responses:
        '200':
          description: Successful delete
        "404":
          description: Not found
  /network/unmanaged:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigUnmanagedListGet
      summary: |
        Get All Unmanaged
      description: |
        Returns the array of the resources that are not administrated by netconfd
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/unmanaged'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigUnmanagedCreate
      summary: |
        Create Unmanaged
      description: |
        Adds an Unmanaged resource. When config is applied, these resources
        and related context are excluded from the applied configuration.
      requestBody:
        description: Unmanaged Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/unmanaged'
      responses:
        '201':
          description: Successful creation
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
  /network/unmanaged/{id}:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigUnmanagedGet
      summary: |
        Get Unmanaged
      description: |
        Retrieve an Unmanaged resource config
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful get
        "404":
          description: Not found
    delete:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigUnmanagedDel
      summary: |
        Delete Unmanaged
      description: |
        Delete Unmanaged resource
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful delete
        "404":
          description: Not found
  /network/rules:
    get:
      x-authorization:
        - support
        - admin
        - monitor
      tags:
        - Network
      operationId: ConfigRulesGet
      summary: |
        Get all ip rules list
      description: |
        Get all ip rules list
      responses:
        '200':
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ruleid'
    post:
      x-authorization:
        - support
        - admin
      tags:
        - Network
      operationId: ConfigRuleCreate
      summary: |
        Configures an IP rule
      description: |
        Configures a L3 IP rule on the main routing table
      requestBody:
        description: Rule Config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/rule'
      responses:
        '201':
          description: Successful creation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ruleid'
        '409':
          description: 
            Resource already exists, or request is incompatible with
            current network state
        "400":
          description:
            Resource config is semantically or syntactically incorrect
...
components:
  schemas:
...
    link:
      type: object
      description: |
        Definition of an ip link interface (emulating iproute2 API)

        References: 
        * [Linux Networking](https://www.kernel.org/doc/Documentation/networking/)
        * [Linux Bonding](https://www.kernel.org/doc/Documentation/networking/bonding.txt)
        * [Links Operstate](https://www.kernel.org/doc/Documentation/networking/operstates.txt)
      properties:
        ifindex:
          description: |
            Inteface index ID
          type: integer
          readOnly: true
        ifname:
          $ref: '#/components/schemas/ifname'
        link:
          allOf:
            - $ref: '#/components/schemas/ifname'
          description: |
            Specify what is the VLAN physical device the virtual device is linked to.
            Applies to vlan type virtual devices
          example: eth0
        flags:
          description: Composition of flags of the interface
          type: array
          items:
            $ref: '#/components/schemas/link_flag'
        mtu:
          description: |
            Maximum Transfer Unit value
          type: integer
        promiscuity:
          description: Promiscuous mode flag
          type: integer
        master:
          allOf:
            - $ref: '#/components/schemas/ifname'
          description: |
            In case the interface is part of a bond or bridge, specifies
            the bond/bridge interface it belongs to.
          example: bond0
        linkinfo:
          oneOf:
          - $ref: '#/components/schemas/linkinfo_device'
          - $ref: '#/components/schemas/linkinfo_dummy'
          - $ref: '#/components/schemas/linkinfo_bridge'
          - $ref: '#/components/schemas/linkinfo_bond'
          - $ref: '#/components/schemas/linkinfo_vlan'
          - $ref: '#/components/schemas/linkinfo_veth'
          - $ref: '#/components/schemas/linkinfo_macvlan'
          - $ref: '#/components/schemas/linkinfo_ipvlan'
          - $ref: '#/components/schemas/linkinfo_ppp'
          - $ref: '#/components/schemas/linkinfo_tun'
          - $ref: '#/components/schemas/linkinfo_tap'
          - $ref: '#/components/schemas/linkinfo_gre'
          - $ref: '#/components/schemas/linkinfo_vrf'
          discriminator:
            propertyName: info_kind
            mapping:
              device: '#/components/schemas/linkinfo_device'
              dummy: '#/components/schemas/linkinfo_dummy'
              bridge: '#/components/schemas/linkinfo_bridge'
              bond: '#/components/schemas/linkinfo_bond'
              vlan: '#/components/schemas/linkinfo_vlan'
              veth: '#/components/schemas/linkinfo_veth'
              macvlan: '#/components/schemas/linkinfo_macvlan'
              ipvlan: '#/components/schemas/linkinfo_ipvlan'
              ppp: '#/components/schemas/linkinfo_ppp'
              tun: '#/components/schemas/linkinfo_tun'
              tap: '#/components/schemas/linkinfo_tap'
              gre: '#/components/schemas/linkinfo_gre'
              vrf: '#/components/schemas/linkinfo_vrf'
        #qdisc:
        #  type: string
        operstate:
          description: |
            Readonly state of the interface.

            Provides information on the state being for example UP of an interface.
            
            It is ignored when applying the config

            Possible values:

              * `up` - Interface is up
              * `down` - Interface is down
              * `unknown` - Interface state is not known (driver does not implement the API)
          type: string
          readOnly: true
          enum:
            - up
            - down
            - unknown
          example: "down"
        #txqlen:
        #  type: integer
        linkmode:
          type: string
          example: "DEFAULT"
        group:
          type: string
          example: "default"
        link_type:
          type: string
          enum:
            - ether
            - loopback
            - ppp
            - gre
            - none
        address:
          type: string
          description: |
            Address on which the link IP layer is based.
            
            This is usually a MAC address, but for some tunnelling protocols virtual
            link interface types (e.g. GRE tunnels) it can also be an IP address of the tunnel. 
          readOnly: true
          example: "08:00:27:b2:9a:ce"

        #broadcast:
        #  type: string
        addr_info:
          type: array
          description: "IP addresses assigned to the link interface"
          items:
            type: object
            properties:
              #family:
              #  type: string
              local:
                allOf:
                - $ref: '#/components/schemas/ip'
                description: IP address assigned to the link interface
              prefixlen:
                type: integer
                description: Network prefix (0-32 for IPv4, 0-128 for IPv6)
                minimum: 0
                maximum: 128
                example: 24
              broadcast:
                type: string
                example: "10.1.2.255"
              address:
                allOf:
                - $ref: '#/components/schemas/ip'
                description: The terminating remote peer for point-to-point kind virtual devices
              #scope:
              #  $ref: '#/components/schemas/scope'
              #dynamic:
              #  type: boolean
              #noprefixroute:
              #  type: boolean
              #label:
              #  type: string
              #valid_life_time:
              #  type: integer
              #preferred_life_time:
              #  type: integer
            required:
              #- family
              - local
              - prefixlen
              #- broadcast
              #- scope
              #- dynamic
              #- noprefixroute
              #- label
              #- valid_life_time
              #- preferred_life_time
      required:
        #- ifindex
        - ifname
        #- flags
        #- mtu
        #- qdisc
        #- operstate
        #- group
        #- txqlen
        - linkinfo
        - link_type
        #- address
        #- broadcast
        #- addr_info
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
# 2) Apply the security globally to all operations
security:
  - bearerAuth: []         # use the same name as above